Add device specific toggling actions
