+++ android/app/build.gradle
@@ -223,6 +223,7 @@ android {
 }
 
 dependencies {
+    implementation project(':react-native-device-settings')
     // loading html files
     compile project(':react-native-local-resource')
     implementation project(':react-native-reanimated')
